update delta transformations without loading_model in jqw alignment

+++ libs/bcg_library/geometry/correspondences/bcg_correspondences.cpp
@@ -104,16 +104,16 @@ MatrixS<-1, -1> correspondences::dense_matrix(size_t M, size_t N) const {
 }
 
 MatrixS<-1, 3> correspondences::get_source(property<VectorS<3>, 3> property) const {
-    MatrixS<-1, 3> P(MatrixS<-1, 3>::Zero(mapping.size(), 3));
-    for (long i = 0; i < mapping.size(); ++i) {
+    MatrixS<-1, 3> P(MatrixS<-1, 3>::Zero(static_cast<long>(mapping.size()), 3));
+    for (long i = 0; i < static_cast<long>(mapping.size()); ++i) {
         P.row(i) = property[mapping[i].row()];
     }
     return P;
 }
 
 MatrixS<-1, 3> correspondences::get_target(property<VectorS<3>, 3> property) const {
-    MatrixS<-1, 3> P(MatrixS<-1, 3>::Zero(mapping.size(), 3));
-    for (long i = 0; i < mapping.size(); ++i) {
+    MatrixS<-1, 3> P(MatrixS<-1, 3>::Zero(static_cast<long>(mapping.size()), 3));
+    for (long i = 0; i < static_cast<long>(mapping.size()); ++i) {
         P.row(i) = property[mapping[i].col()];
     }
     return P;
